requirements.txt
asgiref
Django
djangorestframework
setuptools
sqlparse
typing_extensions
wheel
psycopg2-binary
